# app/openapi.json
{"openapi":"3.1.0","info":{"title":"DRL API — Sổ tay hoạt động sinh viên","description":"Điểm danh chống gian lận, chấm điểm rèn luyện, neo dữ liệu lên Polygon Amoy.","version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Events","description":"Sự kiện và hoạt động"},{"name":"Organizations","description":"Tổ chức cấp phát: trường, khoa, đoàn, CLB…"},{"name":"Auth","description":"Đăng nhập, làm mới token"}],"paths":{"/api/organizations/{id}":{"get":{"tags":["Organizations"],"operationId":"get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrgResponse"}}}}}},"put":{"tags":["Organizations"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrgRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrgResponse"}}}}}}},"/api/events/{id}":{"get":{"tags":["Events"],"operationId":"get_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EventResponse"}}}}}},"put":{"tags":["Events"],"operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EventResponse"}}}}}}},"/api/organizations":{"get":{"tags":["Organizations"],"summary":"Danh sách tổ chức, lọc theo loại","operationId":"list","parameters":[{"name":"type","in":"query","required":false,"schema":{"type":"string","enum":["TRUONG","KHOA","DOAN","HOI_SV","CLB","TRUNG_TAM","DOANH_NGHIEP"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrgResponse"}}}}}}},"post":{"tags":["Organizations"],"summary":"Tạo tổ chức mới","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrgRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrgResponse"}}}}}}},"/api/events":{"get":{"tags":["Events"],"summary":"Danh sách sự kiện, phân trang","operationId":"list_1","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["DRAFT","OPEN","RUNNING","CLOSED"]}},{"name":"orgId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageEventResponse"}}}}}},"post":{"tags":["Events"],"summary":"Tạo sự kiện. Secret HMAC sinh phía server, không nhận từ client.","operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EventResponse"}}}}}}},"/api/auth/refresh":{"post":{"tags":["Auth"],"summary":"Đổi refresh token lấy access token mới","operationId":"refresh","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}}}}},"/api/auth/login":{"post":{"tags":["Auth"],"summary":"Đăng nhập bằng username + password","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}}}}},"/api/events/{id}/status":{"patch":{"tags":["Events"],"summary":"Đổi trạng thái sự kiện","operationId":"changeStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"status","in":"query","required":true,"schema":{"type":"string","enum":["DRAFT","OPEN","RUNNING","CLOSED"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EventResponse"}}}}}}},"/api/auth/me":{"get":{"tags":["Auth"],"summary":"Thông tin tài khoản đang đăng nhập","operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MeResponse"}}}}}}}},"components":{"schemas":{"OrgRequest":{"type":"object","properties":{"name":{"type":"string","minLength":1},"type":{"type":"string","enum":["TRUONG","KHOA","DOAN","HOI_SV","CLB","TRUNG_TAM","DOANH_NGHIEP"]},"parentId":{"type":"integer","format":"int64"},"issuerAddress":{"type":"string","pattern":"^0x[0-9a-fA-F]{40}$"}},"required":["type"]},"OrgResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"type":{"type":"string","enum":["TRUONG","KHOA","DOAN","HOI_SV","CLB","TRUNG_TAM","DOANH_NGHIEP"]},"parentId":{"type":"integer","format":"int64"},"issuerAddress":{"type":"string"},"onChainRegisteredAt":{"type":"string","format":"date-time"}}},"EventRequest":{"type":"object","properties":{"orgId":{"type":"integer","format":"int64"},"title":{"type":"string","minLength":1},"type":{"type":"string","minLength":1},"startAt":{"type":"string","format":"date-time"},"endAt":{"type":"string","format":"date-time"},"location":{"type":"string"},"lat":{"type":"number"},"lng":{"type":"number"},"radiusM":{"type":"integer","format":"int32"},"capacity":{"type":"integer","format":"int32"},"criteriaCode":{"type":"string","pattern":"^C[1-5]$"},"points":{"type":"integer","format":"int32"}},"required":["endAt","orgId","startAt"]},"EventResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"orgId":{"type":"integer","format":"int64"},"title":{"type":"string"},"type":{"type":"string"},"startAt":{"type":"string","format":"date-time"},"endAt":{"type":"string","format":"date-time"},"location":{"type":"string"},"lat":{"type":"number"},"lng":{"type":"number"},"radiusM":{"type":"integer","format":"int32"},"capacity":{"type":"integer","format":"int32"},"criteriaCode":{"type":"string"},"points":{"type":"integer","format":"int32"},"status":{"type":"string","enum":["DRAFT","OPEN","RUNNING","CLOSED"]}}},"RefreshRequest":{"type":"object","properties":{"refreshToken":{"type":"string","minLength":1}}},"TokenResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"tokenType":{"type":"string"},"expiresIn":{"type":"integer","format":"int64"}}},"LoginRequest":{"type":"object","properties":{"username":{"type":"string","minLength":1},"password":{"type":"string","minLength":1}}},"Pageable":{"type":"object","properties":{"page":{"type":"integer","format":"int32","minimum":0},"size":{"type":"integer","format":"int32","minimum":1},"sort":{"type":"array","items":{"type":"string"}}}},"PageEventResponse":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/EventResponse"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"paged":{"type":"boolean"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"}}},"MeResponse":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"username":{"type":"string"},"role":{"type":"string"},"studentId":{"type":"integer","format":"int64"},"mssv":{"type":"string"},"fullName":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}